requirements.txt
# Core dependencies
networkx>=3.0
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12

# Data processing
openpyxl>=3.1
pyyaml>=6.0

# Scientific computing
scikit-learn>=1.3

# Optional but recommended
tqdm>=4.65